Update logic.xml
Removed Exercise 3(c) from Exercises 1.2.1: "I will return the calls and dictate the letter when I arrive at the office." The time aspect introduced by the word "when" made the exercise challenging for students. There are enough other exercises to get at the ideas, so I believe it's best to remove this one.